variables, and Anime and Studio models and add new routes
Extended the Anime model to include studio and audio(referring to new collections), added a new Studio model to define a new database collection for Studios, and included a new environment variable in .env. Also, new routes related to the addition of studios, obtaining genres, and categories were added in the main app file. An existing CRUD operations' logic was revised to accommodate these changes as well.