Add change password route and update User schema
The user routes have been extended by introducing a route to handle change password requests. Another change was adding 'name' field in user creation and 'image' field to the User schema which now has a default value and is required. This allows users to provide a name during registration and includes the name and image in subsequent interactions.