Add collection features to app
Added a new Collection Model to the database and integrated it into the main app. Also, new routes were created to support adding a new collection with title and fetching collections by anime ID, which are populated with the related animes except the one with the ID specified. This added functionality extends the capability of the app to manage and categorize animes in collections.